Move StalledFinishedJob handler to the workers directory.
